delete api folder
