+++ libs/ledger-live-common/src/__tests__/manager.ts
@@ -1,5 +1,5 @@
 import { FirmwareNotRecognized } from "@ledgerhq/errors";
-import Manager from "../api/Manager";
+import Manager from "../manager/api";
 import "./test-helpers/setup";
 describe("getDeviceVersion", () => {
   test("it works with a 1.5.5", async () => {
