--- libs/ledger-live-common/src/api/Fees.ts
@@ -1,30 +1,0 @@
-import invariant from "invariant";
-
-import { FeeEstimationFailed } from "@ledgerhq/errors";
-import { makeLRUCache } from "../cache";
-import { blockchainBaseURL } from "./Ledger";
-import network from "../network";
-import type { CryptoCurrency } from "@ledgerhq/types-cryptoassets";
-
-export type Fees = Record<string, number>;
-
-export const getEstimatedFees: (currency: CryptoCurrency) => Promise<Fees> =
-  makeLRUCache(
-    async (currency) => {
-      const baseURL = blockchainBaseURL(currency);
-      invariant(baseURL, `Fees for ${currency.id} are not supported`);
-      const { data, status } = await network({
-        method: "GET",
-        url: `${baseURL}/fees`,
-      });
-
-      if (data) {
-        return data;
-      }
-
-      throw new FeeEstimationFailed(`FeeEstimationFailed ${status}`, {
-        httpStatus: status,
-      });
-    },
-    (c) => c.id
-  );
