+++ libs/ledger-live-common/src/apps/hw.ts
@@ -26,7 +26,7 @@ import {
   currenciesByMarketcap,
   findCryptoCurrencyById,
 } from "../currencies";
-import ManagerAPI from "../api/Manager";
+import ManagerAPI from "../manager/api";
 import { getEnv } from "../env";
 
 import {
