+++ libs/ledger-live-common/src/hw/installApp.ts
@@ -3,7 +3,7 @@ import { throttleTime, filter, map, catchError } from "rxjs/operators";
 import { ManagerAppDepInstallRequired } from "@ledgerhq/errors";
 import Transport from "@ledgerhq/hw-transport";
 import type { ApplicationVersion, App } from "@ledgerhq/types-live";
-import ManagerAPI from "../api/Manager";
+import ManagerAPI from "../manager/api";
 import { getDependencies } from "../apps/polyfill";
 export default function installApp(
   transport: Transport,
