+++ libs/ledger-live-common/src/hw/installOsuFirmware.ts
@@ -1,7 +1,7 @@
 import Transport from "@ledgerhq/hw-transport";
 import type { OsuFirmware } from "@ledgerhq/types-live";
 import type { Observable } from "rxjs";
-import ManagerAPI from "../api/Manager";
+import ManagerAPI from "../manager/api";
 export default (
   transport: Transport,
   targetId: string | number,
