+++ libs/ledger-live-common/src/hw/uninstallApp.ts
@@ -2,7 +2,7 @@ import { ignoreElements, catchError } from "rxjs/operators";
 import { Observable, throwError } from "rxjs";
 import { ManagerAppDepUninstallRequired } from "@ledgerhq/errors";
 import Transport from "@ledgerhq/hw-transport";
-import ManagerAPI from "../api/Manager";
+import ManagerAPI from "../manager/api";
 import type { App, ApplicationVersion } from "@ledgerhq/types-live";
 export default function uninstallApp(
   transport: Transport,
