+++ libs/ledger-live-common/src/manager/api.ts
@@ -17,13 +17,13 @@ import Transport from "@ledgerhq/hw-transport";
 import { throwError, Observable } from "rxjs";
 import { catchError, map } from "rxjs/operators";
 import { version as livecommonversion } from "../../package.json";
-import { createDeviceSocket } from "./socket";
+import { createDeviceSocket } from "../socket";
 import {
   createMockSocket,
   bulkSocketMock,
   secureChannelMock,
   resultMock,
-} from "./socket.mock";
+} from "../socket/socket.mock";
 import semver from "semver";
 import network from "../network";
 import { getEnv } from "../env";
@@ -43,7 +43,7 @@ import {
   OsuFirmware,
   SocketEvent,
 } from "@ledgerhq/types-live";
-import { getProviderId } from "../manager/provider";
+import { getProviderId } from "./provider";
 
 declare global {
   namespace NodeJS {
