+++ libs/ledger-live-common/src/socket/index.ts
