+++ libs/ledger-live-common/src/socket/socket.mock.ts
